Refactor headers from C to C++ format
This patch simply modifies the coding standard of this repository, from
the C standard of .h to a C++ standard, .hpp. It's relatively painless
to make this change now rather than later.

It's good practice to keep C++ header files distinct from C and may
one-day save a headache when debugging.