CMakeLists.txt copy db into build
For correctness, the database folder structure must be relative to the binary
location, therefore it needs to be copied into the build directory. This patch
makes the minor modification to CMakeLists.txt to do this.

IDEs like Netbeans should be able to directly import and run this project with
minimal hassle.

Ideally no unnecessary copying of resources (potentially hundreds of megabytes
of input images and video files) would be required. It does make developing on
low-spec machines a pain. To that end, in another patch, command line arguments
will be added to support arbitrary paths to reduce the need for hard drive
space and IO on restrictive development platforms like Raspberry Pi.

The db image folder is not version controlled by git, and may never be.  A
subsequent patch may add a link to a ZIP hosted on Dropbox, or another solution
deemed a reasonable compromise. Maybe future unit tests will contain a small
subset of image files.