Logging: Moving from stdout calls to spdlog (C++11) library
System calls are computationally expensive in the main loop so performance
should be improved without them.  This patchset has chosed the spdlog
(github.com/gabime/spdlog) logging framework to replace calls to stdout/stderr
and integer-based log levels

An unfortunate side effect is spdlog requires C++11 standard. This is not a
problem as modern compilers on supported platforms support it, but the CMake
versions on the default software repositories on some currently platforms
(Ubuntu 12.04) are still CMake 2.8, so don't yet support standardized C++
version enforcement of later versions. So currently, we are appending a flag.
This works, but will likely mean compilation on Windows needs slight
modification.