New cropping algorithm - mean shift
In the past, a simple cropping mechanism was used as a proof-of-concept. This
patchset introduces the robust cropping system as described in the original
research paper.

It uses the Mean Shift computer vision algorithm, which finds the area of
highest density in a set of points.

Initially the entire (color classified) image frame is used as input. The from
the centre pixel to every non-background pixel is found. The average of all
points in the image, and the new co-ordinate becomes the new center point. The
frame is slightly cropped and a new center point is found. This is repeated a
number of times until we converge on the center of the image and the frame has
cropped reasonably.

In practise, it produces a very nice crop/normalized image which is ready for
searching with.

In more precise terms, this patch introduces a Mean Shift using a uniform
distribution kernel function, and a variable (monotonically decreasing)
bandwidth parameter

Please note, we have to be careful of the corner cases (literally), currently
we copy in the values to a new matrix to center it.