Formatting all project C++ header and implementation files
This patch uses Netbeans' built in formatter to set codebase-wide standard.

Source files had mixed tabs and spaces (not really noticable if editor not
configured to display whitespace)

This patch standardises on space characters not tab character indentation.
4-space width is used.

Unfortately, a handful of these changes occured in preceding patches, and it's
a bit of pain to retroactively clean up the git history with rebase conflicts
(with comparatively little gain).