Moving image set data to JSON "manifest" files and rapidjson library
This patch refactors the main function, moving the references (filenames/paths)
of data such as lists of training/testing images, into seperate/external
manifest.json files. This allows the main function to be heavily refactored
into a cleaner, more lean function which does not need to be recompiled when
testing different training image sets, for example.

The JSON parsing occurs in the new files manifest.hpp and manifest.cpp.