Axial Mean Shift bandwidth parameter updating and 'precrop' size
Prior to this patch, the mean shift cropping algorithm was using a radius for
the 'window' of considered pixels, known as the bandwidth. While it produced a
reasonable crop, the normalization wasn't centered well.

This patch does bandwidth as a rectangle rather than a circle (more reflective
of the normalized image). Additionally, it is able to change the start and end
points of the x-axis and y-axis independently every iteration of meanshift.

The magnitude of the bandwidth update is determined by the spread of the pixels
in the region of interest (if the non-background pixels are concentrated in one
area, the bandwidth rectangle shrinks faster than if the non-background pixels
take up most of the screen.

Like the original paper, we use standard deviation of the non-background pixels
to do the bandwidth update.

After this patch, normalisation appears to work robustly, whether the mean and
standard deviation of non-background pixels is large or small. In other words,
it's likely a single image can be used to represent a hand pose when the glove
is both close-up and when it is distant.

Additionally, a new command-line argument has been added to allow more precise
cropping by resizing the image after incurring the (large) additional image
processing performance penalties of bilinear filtering and EM classification.
This was initially a seperate patch, but was squashed in with this one and
isn't worth splitting with git diff.