Normalization: Maintain aspect ratio when resizing cropped image
After the meanshift has converged, we're left with a rectangle specifying where
the algorithm thinks the glove is. This rectangle may have arbitrary dimensions
- a tiny horizontal rectangle or a large vertical rectangle.

We need to resize this to the target normalized width/height before we can use
the image as a search query. Unfortunately, the (current) OpenCV API being
targeted, does not have a 'maintain aspect ratio' option in the resize
functions.

This patchset adds the functionality, allowing input video/images to be
normalized in an aesthetically nice way.

Making a rectangular image a square image adds some latency (minute, depending
on how large the widths/heights the user provides) due to processing overhead
(copying data). Additionally, it may or may not add more accuracy to the
database lookup.

Currently only black borders are used (limiting normalized background color to
black). Additionally, the user provided normalization width/heights may not
fully be honored due to squaring the aspect ratio. These issues will be
addressed in a future commit.