Displaying raw input frames for demonstration purposes
Currently, in non-headless mode the design requires the normalized image to be
displayed.

It's trivial to add viewing the raw image stream, functionality which is useful
during demonstration. Unfortunately, displaying a large video file (without
hardware acceleration) slows down the application considerably. Therefore, this
patchset adds a new command-line argument to toggle the raw input image
display.

This patchset adds this functionality to the video stream (not the evaluation
stream yet). Additionally, this patchset leaves command-line argument parsing
in the same manner it was prior to this patch - incomplete support of shortened
arguments which is planned to be addressed/polished after the higher priority
processing pipeline work is completed.