Radically refactor to Google C++ Style Guidelines
gloveTrack is at the point where it needs a 3D rendering backend to produce a
large corpus of synthetic hand poses for a search set, and to render the
closest matches.

It's a good time to clean up the entire application to make integration with
third-party libraries and applications easier, as well as a cleaner API for end
users.