Integrated libhand renderer into gloveTrack
This commit, adds the required files for gloveTrack to render 3D hand poses.
This is intended for construction of a search set of synthetically constructed
hand poses, as well as rendering the closest matches during everyday usage of
the application.

The linking process is currently very much dependent on the folder structure
environment.

The current libhand API isn't very flexible, which forces non-optimal design in
some areas. It's intended gloveTrack project maintain its own fork of libhand
to allow the API flexiblity needed to have a seperate rendering backends such
as Unity, and simply the complex build process of libhand.