Refactors folder structure, adds libhand submodule
Prior to this patchset, gloveTrack required libhand to be downloaded and
compiled seperately.  This is clearly not ideal, as the official libhand v0.9
does not compile on MacOS/Linux distribution without modification.

Additionally, the libhand has some limitations that would be nice to fix and
use as part of gloveTrack.

This patchset tracks a forked version of the most commonly used libhand release
on GitHub as a git submodule.

The REAMDE has been updated to reflect these changes.